edgeai_torchmodelopt/requirements.txt
progressbar
colorama
opencv-python
scipy
matplotlib
pyyaml
torchinfo
